Move inspector related functionality to separate module
